Move next/prev chapters to the Chapter model
