+++ apps/documents/views.py
@@ -1,4 +1,3 @@
-from django.utils.functional import cached_property
 from django.views import generic
 
 from adhocracy4.modules import views as module_views
@@ -73,20 +72,6 @@ class ChapterDetailView(rules_mixins.PermissionRequiredMixin,
     @property
     def chapter_list(self):
         return models.Chapter.objects.filter(module=self.module)
-
-    @cached_property
-    def prev(self):
-        return self.chapter_list\
-            .filter(weight__lt=self.object.weight)\
-            .order_by('-weight')\
-            .first()
-
-    @cached_property
-    def next(self):
-        return self.chapter_list\
-            .filter(weight__gt=self.object.weight)\
-            .order_by('weight')\
-            .first()
 
 
 class DocumentDetailView(ProjectMixin, ChapterDetailView):
